Move testdata/ to repository root
