+++ testdata/large-file.json.gz
